open-world-agents/projects/owa-core/pyproject.toml
[project]
name = "owa-core"
version = "0.5.3"
description = "Core framework and registry system for Open World Agents"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "av>=14.4.0",
    "loguru>=0.7.3",
    "pydantic>=2.10.6",
    "importlib-metadata>=4.0.0; python_version < '3.10'",
    "griffe>=1.7.3",
    "pyyaml>=6.0.2",
    "pillow>=10.0.0",
    "requests>=2.25.0",
    "opencv-python>=4.11.0",
]

# Entry point declaration for built-in std plugin
[project.entry-points."owa.env.plugins"]
std = "owa.env.plugins.std:plugin_spec"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["owa"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "network: marks tests as requiring network access (may be slow or unreliable)",
]
